% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_state_report_text.R
\name{generate_state_report_text}
\alias{generate_state_report_text}
\title{generate_state_report_text}
\usage{
generate_state_report_text(target_state, group_name, df)
}
\arguments{
\item{target_state}{a string.  The name of the state to process}

\item{group_name}{a string.  The name of the taxonomic group to process.}

\item{df}{a tibble.  Summary results from the process_bad_neighbor_group function.}
}
\value{
a string. A paragraph of formatted text.
}
\description{
This is a sample function for generating summary statistics and text for bad neighbor results.
}
\examples{
# load the saved file
comb_all_results_insects <- readr::read_csv("result_csv/insects_comb_result.csv")

generate_state_report_text("Colorado", group_name = "Insect", comb_all_results_insects)
}
