% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_hierarchy_string.R
\name{get_hierarchy_string}
\alias{get_hierarchy_string}
\title{get_hierarchy_string}
\usage{
get_hierarchy_string(taxon_name)
}
\arguments{
\item{taxon_name}{A string for a scientific name for a taxonomic group to search}
}
\value{
A tibble with the returned tsn, taxonomic rank and formatted hierarchy string to use in other parts of the BAP
}
\description{
This a helper function to generate a formatted hierarchy string for a given taxon to use in the Bad Neighbor BAP.  The function uses 
the SOLR service from the Integrated Taxonomic Information System ([ITIS](https://www.itis.gov)).
}
\examples{
# query ITIS to get the string
h_string <- get_hierarchy_string("Poaceae")

# print the result
h_string
# A tibble: 1 x 4
# tsn   rank   taxon   hierarchy_homonym_string
# <chr> <chr>  <chr>   <chr>                   
#    1 40351 Family Poaceae "*\\\\-40351\\\\-*"
    
# use the result in the main bad neighbor query using state name and optional flag
buff_nn <- state_bad_neighbor_query(state_name = "Virginia", taxon = h_string$hierarchy_homonym_string, get_buffer_fips = TRUE)
}
