% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_bad_neighbor_groups.R
\name{process_bad_neighbor_groups}
\alias{process_bad_neighbor_groups}
\title{process_bad_neighbor_groups}
\usage{
process_bad_neighbor_groups(group_name, taxa)
}
\arguments{
\item{group_name}{A string. The informal name of the taxon group to process (e.g. "Tree/Shrub")}

\item{taxa}{A tibble. A table with a the name of the taxon to process along with the associated heirarchy homonym TSN string. The tibbles are typically loded from lookup tables in the "data" directory (see example). The heirarchy homonym TSN string can also be obtained from the [ITIS](https://www.itis.gov) website by searching for the taxon.}
}
\value{
A list of tibbles (data frames) containing the summary results for the Bad Neighbor Analysis from the export_bad_neighbor_list() function, and a summary tibble for all species currently documented in states from the state_all_species_query() function.
}
\description{
This function is a wrapper function executing all necessary steps in generating the lists of bad neighbors.
}
\examples{

# load lookup tables
taxa <- readr::read_csv("data/tree_shrub_heirarchy_strings.csv")

trees_results <- process_bad_neighbor_groups("tree_shrub", taxa = taxa)

}
