% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/state_all_species_query.R
\name{state_all_species_query}
\alias{state_all_species_query}
\title{state_all_species_query}
\usage{
state_all_species_query(fips_list, taxon, state_name)
}
\arguments{
\item{fips_list}{a single state two-digit FIPS code or a parenthetical group of FIPS codes.  Pairs of state FIPS codes and surrounding states are pre-developed and stored in `data/state_lookup.csv`.  Note: the list includes the Distict of Columbia;}

\item{taxon}{a string hierarchy_homonym_string in the form: "*\\-179913\\-*".  Again, a predefined list is stored in: `data/heirarchy_strings.csv`.  The list of taxa is currently restricted to the best represented taxa in BISON.}

\item{state_name}{a string name of the state to process.}
}
\value{
a tibble with the name of the state, the overall occurrence count for all species in the taxon, and a count of the number of species in the taxon in the state.
}
\description{
A BISON query/parsing function returning all species in a given state.  This is included for possible future work, but the results are not used here;
}
\examples{
all_species <- state_all_species_query(fips_list = 51, taxon = "*\\\\-179913\\\\-*", state_name = "Virginia")
}
