% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/state_fips_lookup.R
\name{state_fips_lookup}
\alias{state_fips_lookup}
\title{state_fips_lookup}
\usage{
state_fips_lookup(state_string, fips = NULL)
}
\arguments{
\item{state_string}{a string with the name of the state to search.  This should be the full name of the state
to avoid ambiguous results.}

\item{fips}{(Optional) a 2-digit FIPS code to search}
}
\value{
A tibble with the state FIPS code, the FIPS codes for the states sharing a boundary, and
  the name of the state.
}
\description{
A helper function to find FIPS codes and buffers using state name. The function can also works
in reverse to find state names and FIPS buffers using the state FIPS code.
}
\examples{
# look up the data for Maryland
state_fips_lookup(state_string = "Maryland")

# the lookup is case insensitive 
state_fips_lookup(state_string = "maryland")
state_fips_lookup(state_string = "new york")

# the function will execute partial matches, but multiple results are possible.
# Multiple results are not necessarily useful
state_fips_lookup(state_string = "carolina")

# better to use full names
state_fips_lookup(state_string = "North Carolina")

# look up by FIPS code as string
state_fips_lookup(fips = "02")

# the function will search single number FIPS
state_fips_lookup(fips = 1)
}
