% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_bn_count_map.R
\name{generate_bn_count_map}
\alias{generate_bn_count_map}
\title{generate_bn_count_map}
\usage{
generate_bn_count_map(df, taxon)
}
\arguments{
\item{df}{a tibble.  Typically the results from the process_bad_neighbor_groups function}

\item{taxon}{a string.  The [often informal] name of the taxon group to process (e.g. "Tree/Shrub")}
}
\value{
a ggplot object with a chloropleth map of the bad neighbor counts for states.
}
\description{
A function to generate a chloropleth map of results.
}
\examples{
comb_all_results_forbs \%>\% 
   group_by(state_name) \%>\% 
   summarize(species_count = sum(bad_neighbor_count)) \%>\% 
   # pass to the plot function
   generate_bn_count_map(taxon = "Tree/Shrub")
}
